refactor: RacingGame turn 정보 클래스 생성 및 분리
