+++ README.md
@@ -30,6 +30,7 @@
     - [x] 자동차 이름 정보 클래스 생성
   - [x] 이름 글자 제한 유효성 검사 추가
   - [x] view 메서드 static으로 변경
+  - [x] 레이싱 게임 턴 정보 클래스 생성 및 분리
   
 - 테스트
   - [x] 이름 글자 수 초과에 대해 테스트 코드 추가
