+++ src/main/java/racingcar/domain/game/RacingGame.java
@@ -1,44 +1,37 @@
-package racingcar.domain;
+package racingcar.domain.game;
 
+import racingcar.domain.car.RacingCar;
 import racingcar.domain.power.PowerGeneratable;
 import racingcar.domain.power.RandomPowerGenerator;
 
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
 public class RacingGame {
     private List<RacingCar> cars;
-    private final int finalTurn;
-    private int turnCount;
+    private RacingGameTurnInfo turnInfo;
 
     public RacingGame(List<String> names, int finalTurn) {
         this.cars = names.stream()
                 .map(RacingCar::new)
                 .collect(Collectors.toList());
-        this.finalTurn = finalTurn;
+        this.turnInfo = new RacingGameTurnInfo(finalTurn);
     }
 
     public List<RacingCar> getCars() {
         return this.cars;
     }
 
-    public int getTurnCount() {
-        return this.turnCount;
+    public RacingGameTurnInfo getTurnInfo() {
+        return this.turnInfo;
     }
-
-    public boolean isFinished() {
-        return this.turnCount >= this.finalTurn;
-    }
-
 
     public List<RacingCar> proceedTurn() {
         PowerGeneratable powerGenerator = new RandomPowerGenerator();
         for (RacingCar rc : cars) {
             rc.move(powerGenerator);
         }
-        this.turnCount++;
+        this.turnInfo.increaseTurnCount();
         return this.cars;
     }
