+++ src/main/java/racingcar/view/OutputView.java
@@ -1,6 +1,6 @@
 package racingcar.view;
 
-import racingcar.domain.RacingCar;
+import racingcar.domain.car.RacingCar;
 
 import java.util.ArrayList;
 import java.util.List;
