+++ src/test/java/racingcar/RacingCarNameTest.java
@@ -3,7 +3,7 @@ package racingcar;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import racingcar.domain.RacingCarName;
+import racingcar.domain.car.RacingCarName;
 
 import static org.junit.jupiter.api.Assertions.*;
