+++ src/test/java/racingcar/RacingCarTest.java
@@ -1,10 +1,9 @@
 package racingcar;
 
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import racingcar.domain.RacingCar;
+import racingcar.domain.car.RacingCar;
 
 import static org.junit.jupiter.api.Assertions.*;
